bug fixes and test updates
package wide fixing of error messages and their associated tests
calculate_resampled_wads() problem that would show up when running calculate_Z(). In some low coverage features it is possible that after resampling it would be entirely NA across all sources, leading to a WAD not being able to be calculated and therefore Z not being calculated. Now, calculate_resampled_wads() will return a stop error if any rows are entirely NA, with a message that the filtering criteria should be more stringent